sw/requirements.txt
pymlab
gpsd-py3
numpy
